requirements.txt
ansible==2.2.0.0
cffi==1.9.1
colorama==0.3.7
configobj==5.0.6
cryptography==1.7.1
idna==2.1
Jinja2==2.8
MarkupSafe==0.23
paramiko==2.1.1
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pyfiglet==0.7.5
python-nmap==0.6.1
PyYAML==3.11
six==1.10.0
termcolor==1.1.0
